+++ blocks/show-hide-group/front-end.js
