+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "show-hide-section",
   "version": "1.1.0",
-  "description": "An interactive heading that enables users to reveal or hide its associated content",
+  "description": "Creates accessible summaries that can be toggled to show or hide additional details.",
   "author": "Happy Prime",
   "license": "GPL-2.0-or-later",
   "devDependencies": {
@@ -16,12 +16,12 @@
     "stylelint": "^14.3.0"
   },
   "scripts": {
-    "start": "wp-scripts start",
-    "build": "npm run build:js && npm run build:css",
-    "build:js": "wp-scripts build src/*.js --output-path=build/js/",
-    "lint:js": "wp-scripts lint-js ./src",
-    "fix:js": "wp-scripts lint-js ./src --fix",
-    "build:css": "postcss -o build/css/style.css src/style.css",
-    "lint:css": "stylelint ./src/style.css"
+    "build": "wp-scripts build",
+    "check-engines": "wp-scripts check-engines",
+    "check-licenses": "wp-scripts check-licenses --production",
+    "lint:css": "wp-scripts lint-style '**/*.css'",
+    "lint:js": "wp-scripts lint-js 'blocks/**/*.js'",
+    "lint:fix": "wp-scripts lint-js 'blocks/**/*.js' --fix",
+    "lint:pkg-json": "wp-scripts lint-pkg-json ."
   }
 }
