+++ plugin.php
@@ -23,26 +23,4 @@ if ( ! defined( 'WPINC' ) ) {
 	die;
 }
 
-require_once __DIR__ . '/includes/block-show-hide-group.php';
-require_once __DIR__ . '/includes/block-show-hide-section.php';
-
-add_action( 'enqueue_block_assets', 'hp_show_assets' );
-
-/**
- * Enqueue frontend assets in the document footer.
- */
-function hp_show_assets() {
-	if ( ! has_block( 'happyprime/show-hide-group' ) || is_admin() ) {
-		return;
-	}
-
-	$asset_data = require_once __DIR__ . '/build/js/front-end.asset.php';
-
-	wp_enqueue_script(
-		'hp-show',
-		plugin_dir_url( __FILE__ ) . 'build/js/front-end.js',
-		$asset_data['dependencies'],
-		$asset_data['version'],
-		true
-	);
-}
+require_once __DIR__ . '/includes/block.php';
