--- src/index.js
@@ -1,2 +1,0 @@
-import './blocks/show-hide-group';
-import './blocks/show-hide-section';
